fix: disable `ctrl` server by default
Disable GQL server for chaosctl by default, and only add `pods/exec`
permissions if GQL server is specifically enabled.

This reduces the risk associated with a default install.

The longterm solution could be to make the RPC calls required for the
GQL API to function without using `pods/exec` calls.

Co-authored-by: Andrius Navasaitis <andrius.navasaitis@form3.tech>
Signed-off-by: Graham Brereton <graham.brereton@form3.tech>